Fix buyOrder format suggestion for OneClick.
The suggestion of making the buyOrder a timestamp plus a serial number
is terrible for distributed systems. It turned out that Transbank never
really requires nor validates this format. My bet is that someone put it
into the docs with the best intpention of guiding integrators, but it
turned out to be bad guidance.

The fix is to just remove the suggestion and leave the "unique
identifier" description which for competent integrators should be
enough guidance (and it's their business to figure out the best way to
create unique identifier in their system).